move ReportingAlreadyAttachedException
