--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/exceptions/ReportingAlreadyAttachedException.kt
@@ -1,4 +1,0 @@
-package fr.gouv.cacem.monitorenv.infrastructure.database.repositories.exceptions
-
-/** Thrown when attempting to archive an entity linked to non-archived child(ren). */
-class ReportingAlreadyAttachedException(message: String) : RuntimeException(message)
